dnt/requirements_2.txt
channels-redis==2.4.2
daphne==3.0.2
whitenoise==6.4.0